Rename folders: force case change
