+++ src/main/java/com/planner/business/services/ICourseScraper.java
@@ -1,8 +1,8 @@
-package com.planner.business_temp.services_temp;
+package com.planner.business.services;
 
 import java.util.List;
 
-import com.planner.data_temp.entities_temp.Course;
+import com.planner.data.entities.Course;
 
 public interface ICourseScraper {
     public List<Course> scrapeCourses();
