+++ src/main/java/com/planner/business/services/implementation/ChatService.java
@@ -1,12 +1,12 @@
-package com.planner.business_temp.services_temp.implementation_temp;
+package com.planner.business.services.implementation;
 
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
-import com.planner.business_temp.dto.MessageDTO;
-import com.planner.business_temp.dto.SessionRequest;
+import com.planner.business.dto.MessageDTO;
+import com.planner.business.dto.SessionRequest;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
