+++ src/main/java/com/planner/persistence/repository/DegreeRequirementRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.planner.data_temp.entities_temp.DegreeRequirement;
+import com.planner.data.entities.DegreeRequirement;
 
 @Repository
 public interface DegreeRequirementRepository extends JpaRepository<DegreeRequirement, String> {
