+++ src/main/java/com/planner/persistence/services/ICoursePersistenceService.java
@@ -2,7 +2,7 @@ package com.planner.persistence.services;
 
 import java.util.List;
 
-import com.planner.data_temp.entities_temp.Course;
+import com.planner.data.entities.Course;
 import com.planner.persistence.dto.CourseFilter;
 
 public interface ICoursePersistenceService {
